+++ src/onegov.org/onegov/org/assets/js/ace-mode-form.js
@@ -15,7 +15,7 @@ var FormHighlightRules = function() {
             },
             {
                 token: 'support.function',
-                regex: /(___|\*\*\*|\.\.\.|@@@|YYYY.MM.DD|YYYY.MM.DD HH:MM|HH:MM)/
+                regex: /(___|\*\*\*|\.\.\.|@@@|YYYY.MM.DD|YYYY.MM.DD HH:MM|HH:MM|[0-9]+[0-9\.]+[0-9]+)/
             },
             {
                 token: 'keyword',
@@ -36,6 +36,10 @@ var FormHighlightRules = function() {
             {
                 token: 'support.constant',
                 regex: /\[[0-9]+\]/
+            },
+            {
+                token: 'support.type',
+                regex: /\/[^=]+$/
             }
         ]
     };