+++ src/onegov.org/onegov/org/assets/js/code_editor.js
@@ -1,13 +1,11 @@
 $(function() {
     $('textarea[data-editor]').each(function() {
         var textarea = $(this);
-
         var mode = textarea.data('editor');
         var readonly = textarea.is('[readonly]');
         textarea.css('display', 'none');
 
         var outside = $('<div class="code-editor-wrapper">');
-
         var inside = $('<div class="code-editor">');
         inside.position('absolute');
 
@@ -15,6 +13,7 @@ $(function() {
         outside.insertBefore(textarea);
 
         var editor = ace.edit(inside[0]);
+        editor.$blockScrolling = 1;
         editor.setOptions({
             minLines: 10,
             maxLines: 50
@@ -26,7 +25,6 @@ $(function() {
         editor.renderer.setShowGutter(false);
         editor.renderer.setScrollMargin(10, 10, 10, 10);
         editor.getSession().setValue(textarea.val());
-        console.log(mode);
         editor.getSession().setMode("ace/mode/" + mode);
         editor.setTheme("ace/theme/tomorrow");
 
@@ -34,6 +32,7 @@ $(function() {
             inside.toggleClass('focused', true);
             outside.toggleClass('focused', true);
         });
+
         editor.on("blur", function() {
             inside.toggleClass('focused', false);
             outside.toggleClass('focused', false);
@@ -66,5 +65,52 @@ $(function() {
             }
             textarea.val(editor.getSession().getValue());
         });
+
+        if (mode === 'form') {
+            var fnid = 'onInsertSnippet' + Math.floor(Math.random() * 10001);
+            var toolbar = $(
+                '<div class="formcode-ace-editor-toolbar" ' +
+                     'data-source="/formcode-snippets" ' +
+                     'data-target="' + fnid + '">'
+            );
+
+            window[fnid] = function(snippet, title) {
+                var AceRange = ace.require("ace/range").Range;
+                var session = editor.getSession();
+                var selection = editor.getSelectionRange();
+
+                // insert before or after the line?
+                var insert = selection.end.column === 0 && snippet + '\n' || '\n' + snippet;
+
+                // where to insert?
+                var row = editor.getSelectionRange().start.row;
+                var column = selection.end.column !== 0 && Number.MAX_VALUE || 0;
+                var length = session.getLength();
+
+                // only insert after a line that doesn't begin with a space
+                // (skipping over multiline fields)
+                if (column > 0) {
+                    while (row < length && session.getLine(row + 1)[0] === " ") {
+                        row += 1;
+                    }
+                }
+
+                // set the selection to a single character and insert
+                editor.selection.setRange(new AceRange(row, column, row, column));
+                session.insert({row: row, column: column}, insert);
+
+                // select the title of the snippet
+                editor.selection.setRange(
+                    editor.find(title, {
+                        range: new AceRange(row, 0, row, Number.MAX_VALUE)
+                    })
+                );
+
+                editor.focus();
+            };
+
+            toolbar.insertBefore(outside);
+            initFormSnippets(toolbar.get(0));
+        }
     });
 });